Refactor sites dashboard banners manager
