+++ client/sites/components/sites-dashboard-banners/use-a8c-for-agencies-sites-banner.tsx
@@ -1,16 +1,22 @@
 import { Gridicon } from '@automattic/components';
 import { localizeUrl } from '@automattic/i18n-utils';
 import { translate } from 'i18n-calypso';
+import { useSelector } from 'react-redux';
+import { isCardDismissed } from 'calypso/blocks/dismissible-card/selectors';
 import Banner from 'calypso/components/banner';
 
 export function useA8CForAgenciesSitesBanner( { sitesCount }: { sitesCount: number } ) {
 	const id = 'dismissible-card-a8c-for-agencies-sites';
+	const isA8CForAgenciesBannerDismissed = useSelector( isCardDismissed( id ) );
 
 	return {
 		id,
 		shouldShow() {
-			const showA8CForAgenciesBanner = sitesCount >= 5;
-			return showA8CForAgenciesBanner;
+			if ( isA8CForAgenciesBannerDismissed ) {
+				return false;
+			}
+			// Show banner when user has 5 or more sites
+			return sitesCount >= 5;
 		},
 		render() {
 			return (
@@ -24,7 +30,7 @@ export function useA8CForAgenciesSitesBanner( { sitesCount }: { sitesCount: numb
 					description={ translate(
 						"Earn up to 50% revenue share and get volume discounts on WordPress.com hosting when you migrate sites to our platform and promote Automattic's products to clients."
 					) }
-					dismissPreferenceName="dismissible-card-a8c-for-agencies-sites"
+					dismissPreferenceName={ id }
 					event="learn-more"
 					horizontal
 					href={ localizeUrl( 'https://wordpress.com/for-agencies?ref=wpcom-sites-dashboard' ) }
